Use new session cookie name
Toggle seems to have changed the session cookie name. This caused panics
while using this library. The following line in `requestWithLimit` adds
the session cookie regardless of the value. AddCookie panics when adding
a nil value.

     req.AddCookie(c.cookie)

Changed the session cookie name to the new name used by toggle.
Added a guard in the authenticate method that will error when the session
cookie is not found.